pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "kakao-work-bot-py"
version = "0.1.0"
description = "카카오톡 챗봇 with LangChain"
authors = [
    {name = "shjang"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "langgraph (>=0.6.7,<0.7.0)",
    "langchain-openai (>=0.3.33,<0.4.0)",
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "supabase (>=2.20.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
